@vingkan 2021-02-15
Great work, Lynn! This is a big improvement.

Before, per worker:

```
2021-02-15T10:25:33.134911+00:00 app[web.1]: Done in 7.3 secs
```

After, per worker:

```
2021-02-15T21:42:26.499227+00:00 app[web.1]: Done in 2.5 secs
```

2.5 seconds is still quite a bit. My guess is that the new default model takes a longer time to load because it's bigger.

Sure enough, Lynn's two models are the largest. The new weighted model is about 17MB.

```
(venv) vineshs-mbp:haxml vineshkannan$ ls -la models
total 133576
drwxr-xr-x  12 vineshkannan  staff       384 Feb  8 19:33 .
drwxr-xr-x  18 vineshkannan  staff       576 Feb  3 13:23 ..
-rw-r--r--   1 vineshkannan  staff    239186 Dec 30 15:53 demo_DecisionTree.pkl
-rw-r--r--   1 vineshkannan  staff   3857747 Jan 28 00:30 demo_knn5.pkl
-rw-r--r--   1 vineshkannan  staff       805 Jan 28 00:30 demo_logistic_regression.pkl
-rw-r--r--   1 vineshkannan  staff  13505850 Feb  8 19:33 edwin_classic_random_forest_max_depth_12.pkl
-rw-r--r--   1 vineshkannan  staff   2732634 Feb  8 19:33 edwin_classic_random_forest_max_depth_8.pkl
-rw-r--r--   1 vineshkannan  staff  10881852 Feb  8 19:33 edwin_random_forest_max_depth_12.pkl
-rw-r--r--   1 vineshkannan  staff   2483043 Feb  8 19:33 edwin_random_forest_max_depth_8.pkl
-rw-r--r--   1 vineshkannan  staff     74919 Jan 28 00:30 gradientBoost.pkl
-rw-r--r--   1 vineshkannan  staff  16976508 Feb  8 19:33 lynn_random_forest_max_depth_15_both_def_dist.pkl
-rw-r--r--   1 vineshkannan  staff  17601180 Feb  8 19:33 lynn_random_forest_max_depth_15_only_weighted_dist.pkl
```

But, this time to load the model is paid at startup or at first request per model, so this is still a huge improvement! Now our free Heroku server should respond faster when woken up.